requirements.txt
fastapi
uvicorn
httpx
aio-pika
sqlalchemy[asyncio]
asyncpg
pydantic
pydantic-settings
PyYAML
openai
alembic
psycopg2-binary
python-dotenv
redis
tenacity
aiogram
google-auth
google-api-python-client
pandas
openpyxl